pyproject.toml
[project]
name = "praising-chatbot"
version = "0.3.0"
description = "A supportive and encouraging chatbot with FastAPI backend and Gradio frontend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "gradio>=5.49.1",
    "openai>=1.61.1",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
